// MECProject-App/src/app/models/signup.ts
export class signup {

    username: string;
    email: string;
    phone: string;
    password: string;
    shopname: string;
    shopaddress: string;
    bankname: string;
    bankbranch: string;
    accountnumber: string;
    commission: number;
    userId: number;
    status: string;
    cnic: number;
    city: string;
}